Enhance the system to be more efficient
Removed additional reference storing in the code to make it more
efficient. Added min and max temperatures in the script.

Simplified the fan change state method by using the boolean value of the
condition used for the `fans_output.value()` function.

Made the system fully work by investigating the way the double relay
module worked. It actives on low pulses rather than on high pulses.
Thus, inverting the condition of the fan state is needed to achieve the
desired result.